+++ test/test_core/test_constant_arrays.py
@@ -49,6 +49,24 @@ class Dummy(Computation):
         return plan
 
 
+class DummyOuter(Computation):
+
+    def __init__(self, length, arr1, arr2):
+        assert arr1.shape == (length,)
+        assert arr2.shape == (2, length)
+        self._arr1 = arr1
+        self._arr2 = arr2
+        Computation.__init__(self, [
+            Parameter('output', Annotation(Type(numpy.float32, length), 'o')),
+            ])
+
+    def _build_plan(self, plan_factory, device_params, output):
+        plan = plan_factory()
+        dummy = Dummy(self._arr1.shape[0], self._arr1, self._arr2)
+        plan.computation_call(dummy, output)
+        return plan
+
+
 def test_constant_arrays_computation(thr):
 
     N = 200
@@ -57,6 +75,25 @@ def test_constant_arrays_computation(thr):
     ref = (arr1 * (arr2[0] + arr2[1])).astype(numpy.float32)
 
     d = Dummy(N, arr1, arr2).compile(thr)
+    out_dev = thr.empty_like(d.parameter.output)
+    d(out_dev)
+    test = out_dev.get()
+
+    assert diff_is_negligible(test, ref)
+
+
+def test_constant_arrays_computation_nested(thr):
+    """
+    Check that constant arrays from a nested computation are
+    transfered to the outer computation.
+    """
+
+    N = 200
+    arr1 = get_test_array(N, numpy.int32)
+    arr2 = get_test_array((2, N), numpy.float32)
+    ref = (arr1 * (arr2[0] + arr2[1])).astype(numpy.float32)
+
+    d = DummyOuter(N, arr1, arr2).compile(thr)
     out_dev = thr.empty_like(d.parameter.output)
     d(out_dev)
     test = out_dev.get()