+++ doc/source/history.rst
@@ -16,6 +16,8 @@ Release history
 
 * ADDED: a method :py:meth:`~reikna.core.Type.padded` of :py:class:`~reikna.core.Type`.
 
+* FIXED: a bug where a computation with constant arrays could not be called from another computation.
+
 
 0.7.0 (5 Jul 2018)
 ==================